move subscribe and unsubscribe out of server
